fixup generated file
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@51095 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/minidlna/libid3tag-0.15.0b/compat.h
@@ -34,7 +34,7 @@ struct id3_compat {
 };
 
 struct id3_compat const *id3_compat_lookup(register char const *,
-					   register unsigned int);
+					   register size_t);
 
 int id3_compat_fixup(struct id3_tag *);
